+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -148,7 +148,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-03T13:54:09.644Z",
+    "2026-02-04T07:59:35.795Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {